// src/types/catalogs.type.ts
export type Categories = {
    id_cat: number;
    nombre: string;
}

export type TypeMaterial = {
    id_tipo: number;
    nombre: string;
}

export type Languajes = {
    id_idioma: number;
    nombre: string;
    codigo: string;
}

export type Niveles = {
    id_nivel: number;
    nombre: string;
}